CPU speed value should be float type in filtering
Bug: 883214

+++ Server/bkr/server/needpropertyxml.py
@@ -664,7 +664,7 @@ class XmlCpuSpeed(ElementWrapper):
     """
     def filter(self, joins):
         op = self.op_table[self.get_xml_attr('op', unicode, '==')]
-        value = self.get_xml_attr('value', int, None)
+        value = self.get_xml_attr('value', float, None)
         query = None
         if value:
             joins = joins.join(System.cpu)
